cabin data CRUD done

+++ leafy-lodgings/src/services/apiCabins.js
@@ -2,34 +2,47 @@ import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
     const { data, error } = await supabase.from("cabins").select("*");
+
     if (error) {
         console.error(error);
         throw new Error("Cabins could not be loaded");
     }
+
     return data;
 }
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+    const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
     const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll("/", "");
-    const imagePath = `${supabaseUrl}/storage/v1/object/sign/cabin-images/${imageName}`;
+    const imagePath = hasImagePath
+        ? newCabin.image
+        : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-    // https://oxjsltpyeodateohtyia.supabase.co/storage/v1/object/sign/cabin-images/cabin-001.jpg?token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1cmwiOiJjYWJpbi1pbWFnZXMvY2FiaW4tMDAxLmpwZyIsImlhdCI6MTY5MjU2MTg1NSwiZXhwIjoxNzI0MDk3ODU1fQ.9cTNdCWuOP0XPtvqmf1hKeO00gpWDsi7wLaHd5GHrB0&t=2023-08-20T20%3A04%3A14.693Z
+    // 1. Create/edit cabin
+    let query = supabase.from("cabins");
 
-    //1.Create cabin
-    const { data, error } = await supabase
-        .from("cabins")
-        .insert([{ ...newCabin, image: imagePath }]);
+    // A) CREATE
+    if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+    // B) EDIT
+    if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+    const { data, error } = await query.select().single();
+
     if (error) {
         console.error(error);
         throw new Error("Cabin could not be created");
     }
 
-    //2. upload image
+    // 2. Upload image
+    if (hasImagePath) return data;
+
     const { error: storageError } = await supabase.storage
         .from("cabin-images")
         .upload(imageName, newCabin.image);
 
-    //3.Delete the cabin if there was an error uploading the image
+    // 3. Delete the cabin IF there was an error uplaoding image
     if (storageError) {
         await supabase.from("cabins").delete().eq("id", data.id);
         console.error(storageError);
@@ -46,5 +59,6 @@ export async function deleteCabin(id) {
         console.error(error);
         throw new Error("Cabin could not be deleted");
     }
+
     return data;
 }
